refactor: re-utilisation de compareDebit => déplacement dans common
