--- dbmongo/js/public/compareDebit.js
@@ -1,6 +1,0 @@
-function compareDebit (a,b) {
-  "use strict";
-  if (a.numero_historique < b.numero_historique) return -1
-  if (a.numero_historique > b.numero_historique) return 1
-  return 0
-}
